opti: size header prepend don't requires create a new chunk
The first chunk of the list reserve a space with sizeof(size_t) bytes
before payload.
In this case, the ChunkList can push the chunk to free chunk list and
reuse it later.
The first version is prepend a new chunk with sizeof(size header type)
bytes. Current, I deprecate it.
The second version is make the first chunk be special, it has a header
with sizeof(size_t) bytes but the payload is also CHUNK_SIZE, it can't
be push to the free chunks.
In the end, I make the first chunk same with the other chunk with same
maxsize.
In most case, this is correct and more easy to expand and maintain.