fix(TcpClient): Cycle reference in close callback
The close callback hold a duplicate of client, ie. ref count >= 2 when
client is don't hold by user, the client is not be destoryed.

Pass it by weak refernece can solve this problem but introduce another
problom that the `HandleClose()` of connection also requires the
connection is lived. To solve this, increment connection ref-count.